v 0.1.0 b
Различные исправления
Добавления фильтров сигнала
Начало адресации устройств

+++ src/Measurement.h
@@ -8,7 +8,7 @@ namespace IntroSatLib
 	namespace Units
 	{
 		using DistanceAcceleration = float;
-		using DegreeVelocity = int32_t;
+		using DegreeVelocity = float;
 		using Luminance = uint32_t;
 	}
 }
